COMMON: Remove broken opaque key wrap/unwrap code
The code in common code to support key wrapping and unwrapping of
secure keys (having the CKA_IBM_OPAQUE attribute) is broken. To wrap,
it encrypts the secure key blob from the to be wrapped key's CKA_IBM_OPAQUE
attribute as is. For symmetric keys, the key blob is taken as is, for
asymmetric keys, a non-standard private key BER-encoding of the
private key is generated, containing the secure key blob as octet string,
instead of the private key components.

This is of course not what a wrap function should do for a secure key.
It instead should wrap the effective key of the secure key (encoded in a
standard BER encoding), which can only be done inside the HSM.

The only secure key token that uses the common code for wrap/uwrap is the
CCA token, but its key wrap functionality has already been disabled with
commit 3034875bc4f5b91528cc3e19c759852cb5887276 because the way common code
wraps secure keys is broken.

The TPM token declares itself as a clear key token, but uses opaque keys
for RSA anyway. This fails in the wrap/unwrap code anyway, so this change
does not make the TPM token fail more or less.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>