COMMON: Rework secure key token detection
Remove field token_keysize from struct token_specific_struct and
replace it with a new boolean field secure_key_token. Field token_keysize
was mainly used to determine if its a secure key (token_keysize > 0) or
not. This field is of no other use since the last commit, so lets change
it to a boolean field for secure key token detection only.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>