COMMON: Introduce token specific wrap/unwrap functions
This allows a (secure key) token to provide a token specific key wrap
and unwrap function to support secure key wrapping and unwrapping.
Key wrapping must usually happen inside the HSM for secure keys, so
this can not be handled by common code.

The token specific wrap/unwrap function can request the common code
to perform clear key wrapping, if neither the to be wrapped key, nor the
wrapping key is opaque. Similar, for unwrapping, it can request the common
code to perform clear key unwrapping, if neither the unwrapping key, nor
the unwrapped key is opaque. This can be the case for tokens where not all
key types are opaque, but some key types are clear keys (e.g. the TPM
token).

Secure key tokens must provide a token specific wrap and unwrap function,
otherwise CKR_FUNCTION_NOT_SUPPORTED is returned when the common code
is called to perform key wrapping/unwrapping.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>