COMMON: Fix ECPrivateKey decoding to allow padded data
For EC private key unwrapping, the data passed to the decode function
may be padded, and thus be larger than the actual ECPrivateKey ASN1
structure. So look at the data only up to the length in the first SEQUENCE.
Since an EC private key may include an optional public key, we need to
know the actual length to be able to find out of the optional public key
is present or not.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>