Use streaming parser
Restructured the parser as a streaming parser that emits events to an external
processing entity.  With this method we can use the parser also in the
migration tool to find the correct stdll and also update the configuration
file while preserving comments and newlines (tabs and spaces are not
necessarily preserved).

The rework of the parser also updated the grammar to include empty lines and
whole line comments within the slot definition.  This resolves the ugliness

slot 1 {
 description = something # comment allowed here
}

versus

slot 1 {
  description = something
}

Signed-off-by: Juergen Christ <jchrist@linux.ibm.com>