Make pkcstok_migrate use config parser
Now that the configuration parser is more widely usable, don't try to parse
the config by hand.  Instead, just use the config parser to find the correct
slot.

For writing the new configuration, we still use string operations.  This
commit changes the search for the correct slot to a regular expression search
which should fix some bugs reported in #312.  This is deliberately done to
preserve the order of tokens and the comments in the configuration file.

Fixes #312.

Signed-off-by: Juergen Christ <jchrist@linux.ibm.com>