Change parser error handling.
Parser now treats every error as fatal.  Parsing will move to error state and
stay there since no error recovery strategy is defined.  The parser will also
report back the error state to the caller of the parsing function.

Error reporting is now done via another callout.  This allows users to supply
their own error reporting facility.  Two different error sources are
supported:  The main parser if the token sequence does not yield a valid
production, and the parser callout implementations which might reject a given
setting.  In both cases, the parser will end up in an error state.  In the
first case, error messages will be provided by the parser.  In the second
case, the callout implementor is responsible for storing error information
somewhere to report it when the main parser calls the reporting callback.
This callback also includes the line number the error has been detected.  Note
that this number might be off by one if the error is only detected after a
newline-token is already produced.

Also implemented a stack to keep track of strings produced during scanning of
the input.  This stack is cleared either within the parsing rules, or (in case
of errors) when parsing is completed.

Signed-off-by: Juergen Christ <jchrist@linux.ibm.com>