Remove superfluous parser rule
There were too many rules that allowed multiple successive lines of comments
or newlines.  This caused two shift/reduce conflicts in the parser.  By simply
eliminating the main rule we get a grammar that still recognizes the same
language as the previous one.  But this one does not have any shift/reduce
conflicts.

Signed-off-by: Juergen Christ <jchrist@linux.ibm.com>