TESTCASES: Add tests for special attributes
Check that updating attributes of an object is rejected with CKR_ACTION_PROHIBITED
if CKA_MODIFIABLE is FALSE. Also check that CKA_MODIFIABLE can not be
updated via C_SetAttributeValue().

Check that copying of an object is rejected with CKR_ACTION_PROHIBITED if
CKA_COPYABLE is FALSE. Also check that CKA_COPYABLE can be updated to FALSE,
but not back to TRUE, once it was set to FALSE.

Check that destroying of an object is rejected with CKR_ACTION_PROHIBITED if
CKA_DESTROYABLE is FALSE.

Check that CKA_TRUSTED can not be set to TRUE by a user. Only SO can set CKA_TRUSTED
to TRUE.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>