SOFT: Support raw EC Point with CKM_ECDH1_DERIVE
Allow field pPublicData in CK_ECDH1_DERIVE_PARAMS to
specify either the raw or the DER encoded EC Point.

Per PKCS#11 standard: A token MUST be able to accept this
value encoded as a raw octet string (as per section A.5.2
of [ANSI X9.62]). A token MAY, in addition, support accepting
this value as a DER-encoded ECPoint (as per section E.6 of
[ANSI X9.62]) i.e. the same as a CKA_EC_POINT encoding.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>