Support CKA_PUBLIC_KEY_INFO for public and private keys
Extract the public key info (SPKI) from a public or private key
when it is generated, unwrapped, or imported.

Note that it is not always possible to re-construct the public key
info from a private key. In this case the CKA_PUBLIC_KEY_INFO
remains empty, or whatever was specified in the unwrap or object
creation template.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>