Key derivation via CKM_ECDH1_DERIVE: Make CKA_VALUE_LEN optional
PKCS#11 states: 'This mechanism derives a secret value, and truncates
the result according to the CKA_KEY_TYPE attribute of the template and,
if it has one and the key type supports it, the CKA_VALUE_LEN attribute
of the template. (The truncation removes bytes from the leading end of
the secret value.)'

This implies that CKA_VALUE_LEN is optional. If not specified,
the key type determines the derived key size, or a default is used.
When no KDF is specified in the mechanism parameter, the size of the
derived secret is used as the key size, otherwise the digest size of
the KDF is used.

For keys of type CKK_AES the derived key size must be one of the 3 possible
AES key sizes, otherwise the function fails.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>