API: Move certain parameter checking into the token DLLs
The C_Sign/Verify, C_Sign/VerfyUpdate, C_Sign/VerifyFinal are supposed to
terminate the current operation on a error return. The same is true for
C_Encrypt/Decrypt and friends. To be able to terminate the operation, the
parameter checking must be done by the token DLL, and can not be done by
the API layer. The code in the token DLL already contains these parameter
checks anyway.

Only parameters related to the session must still be checked in the API layer.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>