Add token specific set_attribute_values function
When attributes are changed via C_SetAttributeValues, give the token
a chance to react on an attribute modification. The token can, for
example, reject certain modifications, or it can perform additional
steps on an modification of certain attributes. Secure key tokens may
for example update the CKA_IBM_OPAQUE key blob to reflect the attribute
modifications.

The token specific set_attribute_values is also called when an object
is copied via C_CopyObject. While copying an object, attributes can be
modified or added, so this is basically the same as first copying the
object as-is, and then modify its attributes.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>