EP11: Update attributes in opaque key blob on C_SetAttributeValues
When an attribute is modified via C_SetAttributeValues, update the relevant
attributes inside the EP11 key blob also. EP11 supports only certain boolean
attributes to be modified in the key blob. When one of those attributes is
modified, update them in the key blob, and update CKA_IBM_OPAQUE with the
modified key blob.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>