CCA: support import of raw cca tokens with C_ObjectCreate via CKA_IBM_OPAQUE
The secure key blob for all the keys used with the CCA token is
stored in an attribute CKA_IBM_OPAQUE. The content is a so called
secure CCA key token and a description can be found in the
  "Secure Key Solution with the Common Cryptographic
   Architecture Application Programmer's Guide".

There is a requirement to export and import these raw secure key token
objects. Export is simple: Just fetch the value the CKA_IBM_OPAQUE
attribute for a key object and dump it for example into a file.
Importing such a value was until now not supported. This patch
enables the possibility to import such a raw CCA key token object with
the C_CreateObject() pkcs#11 function.

To import an CCA DES key token, use a template with these mandatory
attributes:
  CKA_CLASS = CKO_SECRET_KEY
  CK_KEY_TYPE = CKK_DES
  CKA_VALUE - dummy but needs to have byte size of the clear key (8)
  CKA_IBM_OPAQUE - with the CCA DES key token

to import an CCA DES3 key token, use a template with these mandatory
attributes:
  CKA_CLASS = CKO_SECRET_KEY
  CK_KEY_TYPE = CKK_DES3
  CKA_VALUE - dummy but needs to have byte size of the clear key (24)
  CKA_IBM_OPAQUE - with the CCA DES3 key token

to import an CCA AES data key token, use a template with these mandatory
attributes:
  CKA_CLASS = CKO_SECRET_KEY
  CK_KEY_TYPE = CKK_AES
  CKA_VALUE - dummy but needs to have byte size of the clear key (16,
	      24, or 32)
  CKA_IBM_OPAQUE - with the CCA AES data key token

to import an CCA AES cipher key token, use a template with these mandatory
attributes:
  CKA_CLASS = CKO_SECRET_KEY
  CK_KEY_TYPE = CKK_AES
  CKA_VALUE - dummy but needs to have byte size of the clear key (16,
	      24, or 32)
  CKA_IBM_OPAQUE - with the CCA AES cipher key token

to import an CCA internal RSA private key token, use a template with
these mandatory attributes:
  CKA_CLASS = CKO_PRIVATE_KEY
  CKA_KEY_TYPE = CKK_RSA
  CKA_IBM_OPAQUE - with the CCA RSA private key token

to import an CCA RSA public key token, use a template with these
mandatory attributes:
  CKA_CLASS = CKO_PUBLIC_KEY
  CKA_KEY_TYPE = CKK_RSA
  CKA_IBM_OPAQUE - with the CCA RSA public key token

to import an CCA HMAC key token, use a template with these mandatory
attributes:
  CKA_CLASS = CKO_SECRET_KEY
  CKA_KEY_TYPE = CKK_GENERIC_SECRET
  CKA_VALUE - with a dummy value the size in bytes of the clear key
	      byte size.
  CKA_IBM_OPAQUE with the CCA RSA private key token.

to import an CCA internal EC private key token, use a template with
these mandatory attributes:
  CKA_CLASS = CKO_PRIVATE_KEY
  CKA_KEY_TYPE = CKK_EC
  CKA_IBM_OPAQUE - with the CCA EC private key token

to import an CCA EC public key token, use a template with these
mandatory attributes:
  CKA_CLASS = CKO_PUBLIC_KEY
  CKA_KEY_TYPE = CKK_EC
  CKA_IBM_OPAQUE - with the CCA EC public key token

Signed-off-by: Harald Freudenberger <freude@linux.ibm.com>