EP11: Pass CKA_MODIFIABLE to EP11 when importing keys
Do not exclude CKA_MODIFIABLE when importing EP11 keys, and pass it to
the related EP11 functions, so that CKA_MODIFIABLE is set correctly
in the key blobs. Otherwise CKA_MODIFIABLE might default to something
else than the attribute in the OCK object, and may result in non-modifiable
EP11 key blobs, although the OCK object is modifiable.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>