Don't allow to set CKA_PUBLIC_KEY_INFO for private keys
Per PKCS#11 the CKA_PUBLIC_KEY_INFO attribute: "DER-encoding of
the SubjectPublicKeyInfo for the associated public key (MAY be
empty; DEFAULT derived from the underlying private key data;
MAY be manually set for specific key types; if set; MUST be
consistent with the underlying private key data)"

And further: "If this attribute is supplied as part of a template
for C_CreateObject, C_CopyObject or C_SetAttributeValue for a private
key, the token MUST verify correspondence between the private key
data and the public key data as supplied in CKA_PUBLIC_KEY_INFO. This
can be done either by deriving a public key from the private key and
comparing the values, or by doing a sign and verify operation. If there
is a mismatch, the command SHALL return CKR_ATTRIBUTE_VALUE_INVALID. A
token MAY choose not to support the CKA_PUBLIC_KEY_INFO attribute for
commands which create new private keys. If it does not support the
attribute, the command SHALL return CKR_ATTRIBUTE_TYPE_INVALID."

Since checking for a consistent SubjectPublicKeyInfo when this attribute
is supplied via template is quite difficult, don't allow it to be
specified in commands which create new private keys for now.

When keys are generated, the SubjectPublicKeyInfo is computed
automatically (if possible) and set to CKA_PUBLIC_KEY_INFO during key
creation.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>