travis: Fix SO pin in travis config
When the tokens to test are initialized, the default SO pin (87654321) is
changed to 76543210. Use this new SO pin for PKCS11_SO_PIN, because the
testcases will use the SO pin from PKCS11_SO_PIN later on.

This fixes an testcase error with pkcs11/attributes.c when run in Travis.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>