Fix attribute-array flatten and unflatten
Some ULONG attributes are always stored as 32 bit ULONG, while others are stored
in the length that CK_ULONG is on the systems architecture. I.e. on a 64 bit
system a CK_ULONG is 64 bit in size, while on a 32 bit system it is 32 bit in size.

The same logic must also be used when unflattening attribute-array elements,
when an element is such an attribute that is always stored as 32 bit ULONG.

Also fix an error when flattening an attribute-array element.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>