EP11: Check if curve is supported on EC key import via C_CreateObject
When importing public or private EC keys that uses unsupported curves (i.e.
Edwards or Montgomery curves), it depends on the EP11 firmware level if the
import fails and how it fails. Even if the import works, the imported EC key
would still not be usable with any mechanisms afterwards, if the underlying
firmware does not support the curve.

Check the curve type when importing EC keys in the EP11 token, and allow Edwards
or Montgomery curves only if all configured APQNs support it. Fail with
CKR_CURVE_NOT_SUPPORTED if not.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>