Avoid deadlock in dlclose() after a fork
Calling dlclose() in a atfork handler may cause a deadlock.
dlclose() may itself modify the atfork handler table to remove
any fork handlers that the to be unloaded library has registered.
Since the atfork handler table is currently locked when we are in
an atfork handler, this would produce a deadlock.

Skip the dlclose() if we are in an atfork handler to avoid the deadlock.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>