EP11: Do not logout EP11 session when in child_fork_initializer
When the EP11 token is configure with STRICT_SESSION mode and/or VHSM mode
enabled, the configured EP11 APQNs are logged in by means of m_Login when
a session is opened, and logged out by means of m_Logout when a session is
closed.

When all currently open sessions are closed by C_Finalize running in the
child_fork_initializer, then do not logout the configured EP11 APQNs,
because this would destroy the EP11 session for the fork parent as well.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>