Fix detection of DER encoded or raw ECDH public data (2)
The previous commit did not completely fix the detection of
the ECDH public data format.

When a raw ECPoint starts with 0x04 followed by a byte who's value
is 2 less than the length of the public data, followed by a byte
that is a valid point conversion indicator (e.g. 0x02 - compressed,
0x04 - uncompressed, or 0x06 - hybrid) then it would have still been
detected as an DER encoded EC Point, resulting in a wrong EC Point.

The detection logic is corrected and is now moved into a separate
function, and is used at all 3 places, i.e. in th soft token, the
ICA token, and the EP11 token.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>