[TEST] Fix setup and cleanup of slots
With the addition of lots of slots to our CI the initialization of slots ran
into timeouts in the expect scripts.  Unfortunately these errors were ignored
(expect standard action) which caused uninitialized slots to be tested.
Remove the 5 second timeout in preference of the default 10 seconds and treat
timeout errors in the expect script.

Also provide a cleanup target for the CI that correctly removes all token data
such that the next run starts from a clean state.

Do not attempt to reinitialize slot 42 if it is already initialized.

Switch slot 45 to use PKEYs since the default and disabled setting currently
cover the same code paths.

Signed-off-by: Juergen Christ <jchrist@linux.ibm.com>