COMMON: Prevent unsavable operation state to be exported
Tokens using OpenSSL 3.0 to implement digest operations (SHA, MD5) are no
longer able to store its digest state in the session context in a way
that it could be exported via C_GetOperationState(). OpenSSL 3.0 does not
provide support to get the digest state. A token must therefore place
pointers to OpenSSL digest contexts into the session state structure.
Such a state can not be externalized through C_GetOperationState().

Also see the discussion in OpenSSL issue "Digest State Serialization":
https://github.com/openssl/openssl/issues/14222

Allow a token to mark an operation context as 'not saveable', which will
cause C_GetOperationState() to return CKR_STATE_UNSAVEABLE if it is tried
to save such a state.

Also, such operation contexts can not simply be freed that way the common
code performs that. Allow a token to use a custom context free function,
to cleanup such complex context structures.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>