API: Use own OpenSSL library context for Opencryptoki's use of OpenSSL
Create a separate library context for Opencryptoki's use of OpenSSL services
and explicitly load the 'default' provider for this context.

This prevents call-loops when the calling application has configured a PKCS#11
provider that uses Opencryptoki under the covers. This could produce a loop
with the following calling tree:
  Application -> Openssl -> PKCS11-provider -> Opencryptoki -> OpenSSL
  -> PKCS11-provider -> Opencryptoki -> ...
Explicitly using the 'default' provider only for Opencrypoki's OpenSSL usage
breaks this loop.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>