EP11: Update EP11 specific attributes from key blob
After a key generation, key import, derive, or unwrap operation,
retrieve some EP11 specific attributes from the key blob, and
update/set these attributes in the Opencryptoki object as well.

The EP11 specific attributes can be specified by an application
in the template for object creation, import, derive, or unwrap,
and they will then be contained in the EP11 key blob as well.
With this change, EP11 specific attributes that have not been
specified and got a default value by EP11, are now also available
as PKCS#11 attributes.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>