[COMMON,ICSF] Fix permission of token SHM segment.
The shm token segments were generated with permission 0666 for the user that
created the segment and the corresponding primary group.  This is potentially
a security problem since especially the segmant was word-rw.

This patch changes the permission to 0660 and changes the ownership of the
segment on creation to group "pkcs11".  Since every user that should be
allowed to use OCK has to be in that group, we should not have any usage
regressions.  If the segments already exist, a reboot might be necessary (or
any other way that cleans up the segments).

Signed-off-by: Juergen Christ <jchrist@linux.ibm.com>