CI: Don't leak an EP11 session when setting the VHSM-pin
When setting the VHSM-pin for slot 42, the expect script killed the
pkcsep11_session process on receipt of the "VHSM-pin successfully set."
message due to 'exit 0'.

However, after that message pkcsep11_session still performs some cleanup,
i.e. removing the EP11 session it opened itself (if STRICT mode is enabled,
too). Because the expect script killed the process before it was able to
perform the cleanup, an EP11 session might be leaked.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>