TESTCASES: EC: Check if key is usable after C_DeriveKey()
Perform a HMAC sign operation with the derived key to check hat
this key is really usable (if the derived key is within the allowed
key sizes of the HMAC mechanism).

We recently encountered a problem with EP11 on a CEX5 where a derived
key blob might be invalid for certain key sizes due to a firmware bug.
This test would fail for such situations.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>