API: Retain the OpenSSL error stack state on API calls
Internal use of OpenSSL calls may cause additional errors to be pushed
to the current thread's OpenSSL error stack. This may interfere with
the calling application's use of the OpenSSL error stack, and may leak
Opencryptoki internal error information to the application.

On exit of all Opencryptoki API calls, make sure to restore the current
thread's error stack to the state it was on entry to the Opencryptoki
API call.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>