[COMMON,ICSF] Fix out-of-sourcetree builds.
When building outside the source tree, generated files should not be generated
inside the source tree, but in the separate build tree.  Unfortunately, the
ylwrap utility from the autotools does not correctly do that when called from
autotools generic prefix rule for make.  Write explicit make rules to fix
this.

Signed-off-by: Juergen Christ <jchrist@linux.ibm.com>