CONFIG: Fix makefile
BUILT_SOURCES should not contain the generated headers.  They are generated
and updated via ylwrap which does a content comparison for all but the
generated parser/lexer C sources.  If the inputs to bison or flex change but
still produce the same header file, make install will regenerate the parser
and lexer, recompile them, and relink all binaries that use the config lib.
By using the generated C files in the BUILT_SOURCES variable we solve this
problem since the C files are always updated by ylwrap (even if they did not
change).

Signed-off-by: Juergen Christ <jchrist@linux.ibm.com>