PKCSSLOTD: Update tokversion in example config
The default opencryptoki.conf did not specify the token version and thus
defaulted to the legacy token store.  Update the sample configuration to use
the 3.12 token store (i.e., the FIPS compliant store).

Note: This will not change anything since by default we do not overwrite the
configuration on installation.  But if new installations are based on this
configuration file, they will directly start with the FIPS compliant store.

Signed-off-by: Juergen Christ <jchrist@linux.ibm.com>