DOC: Added small development documentation
What to do when adding a new mechanism or key type to openCryptoki.  This only
captures the part that is needed in the policy code such that the new
mechanism and/or key type is known to the policy.  It does not capture the
code/part that is needed in the tokens to enforce the policy.

Signed-off-by: Juergen Christ <jchrist@linux.ibm.com>