Config: Add bare word constants
The inital configlib commit did not include the possibility to use bare word
constants.  This is needed, e.g., for the "disable-event-support" keyword of
opencryptoki.conf.

To support this, add ConfigNodeBareConst to keep the constant in the key
without any further values in the node.  Also restructure the grammar:
- Every configelem now takes care of the positions of comments _within_ and
_after_ the element.
- Since "after element i" is "before element (i+1)", the config element cannot
take care of comments before the element (shift/reduce conflict since the
parser cannot know if the comment is after element i or before element i+1).
To alleviate this situation and preserve comments at the beginning of a file
or a structure content, add a commentedconfigelemstar that only takes care of
comments before the first configelem.  That should cover all comment
positions.

Signed-off-by: Juergen Christ <jchrist@linux.ibm.com>