Rework part of token initialization
Do not check user/group permissions in every token on token initialization.
The check can easily be done in API library once for all tokens.

Also allocate the TokData structure in the API library.  This allows the API
library to pass other objects to the tokens without changing the signature of
the initialization function.

Signed-off-by: Juergen Christ <jchrist@linux.ibm.com>