Implement and enforce policies
Implement strength and policy parsing and enforce policies in all tokens.  The
main object of the policy is the struct objstrength which carries the strength
of a key along with a flag specifying if the EC is allowed, and a (maximal)
signature length if this key is used to sign or verify and the length of the
signature depends on the key.

A valid strength configuration has to be found.  If no policy configuration is
found, policies are simply deactivated and keys are only classified according
to the strength definition, but can be used without restrictions.  If both, a
strength and a policy configuration, are available, both have to be valid for
OpenCryptoki to start.  If either of these two configurations fails to parse
or validate, OpenCryptoki will fail with a general error since the system is
not in a valid state.

The configuration files have to be owned by root:pkcs11 and have to have modes
0640.  This is done for security reasons to prevent anybody except root from
altering the policy or strength definitions.

Signed-off-by: Juergen Christ <jchrist@linux.ibm.com>