ICSF: replace PBKDF implementation by libcrypto call
The PBKDF2 implementation of the ICSF token is not correct, and thus
can not be simply replaced by the OpenSSL implementation of PBKDF2.

Introduce a version field in ICCF's MK_SO and MK_USER files to be able to
detect old version files. Old version files do not have a version field,
but start with the total-length field right away. The length contained
in the total-length field can be distinguished from a valid version value.
That way we can still read old version files, but can also transparently
support the new version files with the correct PBKDF2 implementation.

When writing ICSF MK_SO or MK_USER files, we always use the new version
format. This allows a smooth transition to the new version. Just set new
pins for SO and USER and the MK files are updated with the new version
format.

Based on the work from Patrick Steuer:
https://github.com/opencryptoki/opencryptoki/pull/318

Suggested-by: Patrick Steuer <patrick.steuer@de.ibm.com>
Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>