COMMON: DH: Do not create default CKA_VALUE_BITS for import and unwrap
CKA_VALUE_BITS must not be specified for C_CreateObject and C_UnwrapKey.
Do not create a default CKA_VALUE_BITS (with value 0) for import and
unwrap either.

Having a zero CKA_VALUE_BITS attribute may cause failures later on in the
token, i.e. for the EP11 token with newer firmware levels.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>