API: Avoid deadlock when stopping event thread
Avoid that the event thread writes trace messages while it is
enabled for thread cancellation. This might leave the trace mutex in
the locked state and cause subsequent trace calls to lock forever
(e.g in stop_event_thread() right after canceling the thread).

Disable cancellation right at the beginning of the thread function,
and disable it before calling a trace function or leaving the loop.

Also make sure that the cleanup handler is registered and the
cancellation type is set before initially enabling cancellation.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>